Extract new appointment form as a component
